% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.cor_table.R
\name{cor_table}
\alias{cor_table}
\title{APA style correlation table}
\usage{
cor_table(
  data,
  filename = NA,
  table.number = NA,
  show.conf.interval = TRUE,
  show.sig.stars = TRUE,
  landscape = TRUE
)
}
\arguments{
\item{data}{Project data frame}

\item{filename}{(optional) Output filename document filename (must end in .rtf or .doc only)}

\item{table.number}{Integer to use in table number output line}

\item{show.conf.interval}{(TRUE/FALSE) Display confidence intervals in table. This argument is deprecated and will be removed from later versions.}

\item{show.sig.stars}{(TRUE/FALSE) Display stars for significance in table.}

\item{landscape}{(TRUE/FALSE) Make RTF file landscape}
}
\value{
APA table object
}
\description{
APA style correlation table
}
\examples{
#forked from https://github.com/dstanley4/apaTables/
}
