% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.adverse_impact.R
\name{f.adverse_impact}
\alias{f.adverse_impact}
\title{Compute adverse impact statistics}
\usage{
f.adverse_impact(
  d.group,
  d.pass,
  majority.group,
  minority.group,
  min_percent_minority = 0.05
)
}
\arguments{
\item{d.group}{is a vector of grouping variable values}

\item{d.pass}{is a vector of outcome values}

\item{majority.group}{is the label of value assigned to the majority group}

\item{min_percent_minority}{is the minimum percent of the sample that must be in the minority group in order to report results}
}
\value{
Returns a dataframe with columns...

total.n = overall sample size

majority.n = sample size of the majority group

minority.n = sample size of the minority group

percent.minority = percentage of sample in minority group

SR.total = selection ratio of the entire sample

SR.majority = selection ratio of the majority group

SR.minority = selection ratio of the minority group

impact.ratio = SR.minority / SR.majority

fishers.p = p value of the Fisher Exact Test

SD2.test = results of the 2 Standard Deviation Test

lower_95CI, upper_95CI = lower and upper confidence intervals around Impact Ratio

shortfall = the number of additional minority group members needed to pass to have an impact ratio = 1
}
\description{
Accepts grouping vector d.group, vector of outcome (e.g., pass/fail) as d.pass, 
and value for majority group (e.g., "male"), assumes only two groups
}
\examples{
require('psych')
pass <- ifelse(sat.act$ACT > mean(sat.act$ACT) , 1, 0)
table(sat.act$gender)
f.adverse_impact(sat.act$gender, pass, majority.group = 2, minority.group = 1)
}
\author{
Adam Meade \email{awmeade@ncsu.edu}
}
