% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.corstar.R
\name{f.corstar}
\alias{f.corstar}
\title{Add asterisk(s) to significant correlations}
\usage{
f.corstar(corr.obj, is.triangle = TRUE, p.val.1 = 0.05, p.val.2 = 0.01)
}
\arguments{
\item{corr.obj}{is the output list object from the corr.test function in the Psych package}

\item{is.triangle}{is whether the correlation is a symetric matrix (default) or a rectangular 
matrix in which one set of variables is correlated with a different set.}

\item{p.val.1}{is the p-value desired to trigger a single astrisk (default = .05)}

\item{p.val.2}{is the p-value desired to trigger a second astrisk added to the first
default = .01. p.val.2 should be a smaller p val than p.val.1. set p.val.2 = FALSE if
no second astrisk is desired}
}
\description{
Uses an output object from the Psych package and adds 1 or 2 askterisks (stars) to the 
correlation for APA table output. Returns object as dataframe.
}
\examples{
R = matrix(cbind(1,.80,.2,.80,1,.7,.2,.7,1),nrow=3)
U = t(chol(R))
set.seed(1)
random.normal = matrix(rnorm(dim(U)[1]*100,0,1), nrow=dim(U)[1], ncol=100);
X = as.data.frame(t(U \%*\% random.normal))
require('psych')
corrs.1 = corr.test(X)
f.corstar(corrs.1)
f.corstar(corrs.1, p.val.1 = .01, p.val.2 = FALSE)
corrs.2 = corr.test(X[1:2],X[3])
f.corstar(corrs.2,is.triangle = FALSE)
}
\author{
Adam Meade \email{awmeade@ncsu.edu}
}

