% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.describe.R
\name{f.describe}
\alias{f.describe}
\title{Simplify output of R Psych package describe function}
\usage{
f.describe(f.d)
}
\arguments{
\item{f.d}{is the dataframe object}
}
\description{
Accepts dataframe as arguement and returns dataframe table
}
\examples{
\dontrun{
require('psych')
 f.describe(sat.act)
 }
}
\author{
Adam Meade \email{awmeade@ncsu.edu}
}
