% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.get.reg.output.R
\name{f.get.reg.output}
\alias{f.get.reg.output}
\title{Take output from lm function and put relevant info into a dataframe}
\usage{
f.get.reg.output(out.lm)
}
\arguments{
\item{out.lm}{results of a linear regression from lm()}
}
\description{
Returns a dataframe with predictors and coefficients listed along with model statistics 
F,df,p, and r-squred values.
}
\examples{
\dontrun{
model.out <- lm(sat.act[,1]~sat.act[,2]+sat.act[,3])
f.get.reg.output(model.out)
}
}
\author{
Adam Meade \email{awmeade@ncsu.edu}
}

