% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.screen.outliers.R
\name{f.screen.outliers}
\alias{f.screen.outliers}
\title{Deletes multivariate outliers using Mahalanobis distance}
\usage{
f.screen.outliers(all.data, screening.vars = NULL, p.val = 0.05)
}
\arguments{
\item{all.data}{is the database from which to remove outliers}

\item{screening.vars}{is an optional array of variable names on which the screening should be based.}

\item{p.val}{is the cutoff value of the chi-square distribution to use. default is .05.}
}
\description{
Accepts a dataframe and optional list of variables within that dataframe for which to screen the data.
The function computes mahalanobis distace and associated chi-square on the screening data and then
returns a dataframe that is a subset of the original all.data dataframe based on non-significant 
chi-square values.
}
\examples{
\dontrun{
nrow(trees)
new.data <- f.screen.outliers(trees)
nrow(new.data)
}
}
\author{
Adam Meade \email{awmeade@ncsu.edu}
}
