% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.t_test.R
\name{f.t_test}
\alias{f.t_test}
\title{Returns the results of a t-test along with a Cohen's D effect size estimate}
\usage{
f.t_test(df1, df2)
}
\arguments{
\item{df1}{is the majority group dataframe}

\item{df2}{is the minority group dataframe}
}
\description{
Requires two dataframes with continuous variables, variance on variables, and matched in variable order. 
Requires use of describe from 'psych' package. Requires use of cohen.d from 'effsize' package
}
\examples{
\dontrun{
 require('psych')
 require('effsize')
 automatic  <- mtcars[which(mtcars$am == 0),]
 manual <- mtcars[which(mtcars$am == 1),]
 f.t_test(automatic,manual)
}
}
\author{
Adam Meade \email{awmeade@ncsu.edu}
}
