% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.get.reg.output.R
\name{f.get.reg.output}
\alias{f.get.reg.output}
\title{Take output from lm function and put relevant info into a dataframe}
\usage{
f.get.reg.output(out.lm)
}
\arguments{
\item{out.lm}{results of a linear regression from lm()}
}
\description{
Returns a dataframe with predictors and coefficients listed along with model statistics 
F,df,p, and r-squred values.
}
\examples{
R = matrix(cbind(1,.80,.2,.80,1,.7,.2,.7,1),nrow=3)
U = t(chol(R))
set.seed(1)
random.normal <- matrix(rnorm(dim(U)[1]*100,0,1), nrow=dim(U)[1], ncol=100);
X <- as.data.frame(t(U \%*\% random.normal))
model.out <- lm(X[,1]~X[,2]+X[,3])
f.get.reg.output(model.out)
}
\author{
Adam Meade \email{awmeade@ncsu.edu}
}

