% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.write.corrs.R
\name{f.write.corrs}
\alias{f.write.corrs}
\title{Formats in APA format and writes correlation matrices to an excel sheet}
\usage{
f.write.corrs(f.name = "output.xlsx", s.name, obj.corrs, print.p = TRUE,
  print.n = TRUE, ...)
}
\arguments{
\item{f.name}{is the file name to which to write the data. default = 'output.xlsx'}

\item{s.name}{is the name of the Excel sheet(s) to which to write the data}

\item{obj.corrs}{is the correlation object from the corr.test function in the Psych package}

\item{print.p}{requests printing of p values. default = TRUE}

\item{print.n}{requests printing of sample size (n). default = TRUE}

\item{...}{parameters passed to the f.corstar function}
}
\description{
Uses an output object from the Psych package and adds 1 or 2 askterisks (stars) to the 
correlation for APA table output. Returns object as dataframe. Writes correlational 
data for r, n, and pvalues to three consecutively named Excel sheets.
}
\examples{
library('psych')
R = matrix(cbind(1,.80,.2,.80,1,.7,.2,.7,1),nrow=3)
U = t(chol(R))
set.seed(1)
random.normal = matrix(rnorm(dim(U)[1]*100,0,1), nrow=dim(U)[1], ncol=100);
X = as.data.frame(t(U \%*\% random.normal))
corrs = corr.test(X)
f.write.corrs(s.name='example',obj.corrs=corrs)
f.write.corrs(f.name='example.xlsx',s.name='example',obj.corrs=corrs,p.val.2=FALSE)
}
\author{
Adam Meade \email{awmeade@ncsu.edu}
}

